+++ src/MappedValues.h
@@ -13,21 +13,22 @@ class Maps
 {
 public:
 
-    static const std::map<TileSize, uint> tileSizeInt;
-    static const std::map<FontSize, uint> fontSizeInt;
-    static const std::map<BoardSize, uint> boardSizeInt;
+    static const std::map<TileSize, size_t> tileSizeInt;
+    static const std::map<FontSize, size_t> fontSizeInt;
+    static const std::map<BoardSize, size_t> boardSizeInt;
     static const std::map<TileColor, const QString> tileColorStyle;
-    static const std::map<int, TileSize> sliderTileSize;
+    static const std::map<size_t, TileSize> sliderTileSize;
     static const std::map<TileSize, FontSize> tileSizeFontSize;
 
     static int getFontSizeInt( TileSize tileSize );
-    static BoardSize getBoardSizeByInt( uint value );
-    static TileSize getTileSizeByInt( uint value );
+    static BoardSize getBoardSizeByInt( size_t value );
+    static TileSize getTileSizeByInt( size_t value );
     static int getSliderByTileSize( TileSize value );
 
 private:
 
-    template<typename T, typename U> static T getMapKeyByValue( const std::map<T, U>&, U  value );
+    template <typename T, typename U> 
+    static T getMapKeyByValue( const std::map<T, U>&, U value );
 };
 
 #endif
